Final Version of the Report of the Challenge.
I had trouble with the visualization of the images and the formulas when I want to see them in the GitHub web version. When I open it in my VScode app the report with the images and the formulas seems ok with an optimal view of the images and the formulas. In order to present all my work in the most clear way I also uploaded the PDF version with all the images and formulas.